+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Changed
+- `dict_to_object` function no longer requires target class to have a parameterless (or with default values) constructor.
+
 ## [1.0.0] - 2025-02-09
 
 ### Added